// src/order/dto/payment-status.dto.ts
class AmountPayment {
  value: string;
  currenct: string;
}

class ObjectPayment {
  id: string;
  status: string;
  payment_method: {
    type: string;
    id: number;
    saved: boolean;
    title: string;
    card: object;
  };
  created_at: string;
  expires_at: string;
  description: string;
}

export class PaymentStatusDto {
  event:
    | 'payment.succeeded'
    | 'payment.waiting_for_capture'
    | 'payment.canceled'
    | 'refund.succeeded';
  type: string;
  object: ObjectPayment;
}